Switch to a per statement OCI error handle.
Using the global one for everything is not recommended by
Oracle as multiple threads might overwrite each other's error
messages.

It also simplifies the API as Envhp was only being passed in
to the statement related calls for access to its errhp

This opens the possibility for a user to open an environment
but not keep it around. An interesting test case here..