Re-thought the relationship between BindVars and Stmt.
Keeping them together became annoying for API calls that just
need a stmthpi (e.g. preparing a new stmt using an existing
Stmt handle).

Now the bind map is passed from Erlang and
each call to OCIBind will return an updated map of refs.

It is up to the user to pass the Bind Map into the correct execute